requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
build==0.10.0
certifi==2022.12.7
cffi==1.15.1
Chameleon==3.10.2
charset-normalizer==3.0.1
check-manifest==0.49
cmarkgfm==2022.10.27
colorama==0.4.6
defusedxml==0.7.1
devpi==2.2.0
devpi-client==6.0.3
devpi-common==3.7.2
devpi-server==6.8.0
devpi-web==4.2.0
Django==4.1.6
django-cors-headers==3.13.0
django-filter==22.1
django-mongodb-engine==0.6.0
djangorestframework==3.14.0
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.3.0
docopt==0.6.2
docutils==0.19
easy-thumbnails==2.8.5
easy-thumbnails-rest==1.1.4
frozenlist==1.3.3
hupper==1.11
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
lazy==1.5
lxml==4.9.2
MouseInfo==0.1.3
multidict==6.0.4
packaging==21.3
passlib==1.7.4
PasteDeploy==3.0.1
pdf2image==1.16.3
pep517==0.13.0
Pillow==9.4.0
pipreqs==0.4.11
pkginfo==1.9.6
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==3.0.0
pluggy==1.0.0
psycopg2-binary==2.9.5
py==1.11.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.14.0
PyJWT==2.6.0
pymongo==3.11.2
PyMsgBox==1.0.9
PyMuPDF==1.21.1
pyparsing==3.0.9
pyPdf==1.13
PyPDF2==3.0.1
pyperclip==1.8.2
pyproject_hooks==1.0.0
pyramid==2.0.1
pyramid-chameleon==0.3
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python-docx==0.8.11
pytweening==1.0.4
pytz==2022.7.1
readme-renderer==37.3
repoze.lru==0.7
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
soupsieve==2.4
sqlparse==0.2.4
strictyaml==1.6.2
tomli==2.0.1
translationstring==1.4
tzdata==2022.7
urllib3==1.26.14
venusian==3.0.0
waitress==2.1.2
Wand==0.6.11
webencodings==0.5.1
WebOb==1.8.7
Whoosh==2.7.4
yarg==0.1.9
yarl==1.8.2
zope.deprecation==4.4.0
zope.interface==5.5.2
